requirements.txt
requests
PyYAML==3.11
--extra-index-url https://pypi.mss.centurylink.com
lambda-base

# Elasticsearch 5.x
elasticsearch>=5.0.0,<6.0.0